have a util class
